requirements.txt
celery==5.5.3
dj-rest-auth==4.0.1
Django==4.2.7
django-allauth==0.50.0
django-cors-headers==4.3.1
django_celery_results==2.6.0
django-filter==24.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dj-rest-auth[with_social]==4.0.1
gunicorn>20.0.0,<21.0
numpy==2.2.3
pandas==2.2.3
pillow==11.0.0
psycopg2==2.9.10
python-dotenv==1.0.0
pytz==2023.3.post1
redis==6.2.0
requests==2.31.0
requests-oauthlib==1.3.1